pyproject.toml
[tool.poetry]
name = "pdf-chat"
version = "0.1.0"
description = ""
authors = ["Thiago Badaró <badarots@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.1"
google-generativeai = "^0.5.2"
numpy = "^1.26.4"
langchain-text-splitters = "^0.0.1"
langchain-community = "^0.0.38"
pypdf = "^4.2.0"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
